rename Devices folder to Providers
