+++ Assets/MixedRealityToolkit.SDK/Profiles/DefaultMixedRealityRegisteredServiceProvidersProfile.asset
@@ -15,7 +15,7 @@ MonoBehaviour:
   isCustomProfile: 0
   configurations:
   - componentType:
-      reference: Microsoft.MixedReality.Toolkit.Core.Devices.UnityInput.UnityJoystickManager,
+      reference: Microsoft.MixedReality.Toolkit.Core.Providers.UnityInput.UnityJoystickManager,
         Microsoft.MixedReality.Toolkit
     componentName: Unity Joystick Manager
     priority: 10
@@ -50,7 +50,7 @@ MonoBehaviour:
     runtimePlatform: 9
     configurationProfile: {fileID: 0}
   - componentType:
-      reference: Microsoft.MixedReality.Toolkit.Core.Devices.UnityInput.UnityTouchDeviceManager,
+      reference: Microsoft.MixedReality.Toolkit.Core.Providers.UnityInput.UnityTouchDeviceManager,
         Microsoft.MixedReality.Toolkit
     componentName: Unity Touch Input Manger
     priority: 10
