--- Assets/MixedRealityToolkit/Attributes/README.md
@@ -1,3 +1,0 @@
-# Mixed Reality Toolkit - Internal - Attributes
-
-This folder contains all the MRTK Internal editor attributes used to customise the Mixed Reality Toolkit experience in the editor.
