--- Assets/MixedRealityToolkit/Attributes/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 98e013c80d944bf4e93d825f93fd3ba3
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
