+++ Assets/MixedRealityToolkit/Definitions/Devices/MixedRealityControllerMapping.cs
@@ -3,8 +3,8 @@
 
 using Microsoft.MixedReality.Toolkit.Core.Attributes;
 using Microsoft.MixedReality.Toolkit.Core.Definitions.Utilities;
-using Microsoft.MixedReality.Toolkit.Core.Devices;
 using Microsoft.MixedReality.Toolkit.Core.Interfaces.Devices;
+using Microsoft.MixedReality.Toolkit.Core.Providers;
 using System;
 using System.Runtime.CompilerServices;
 using UnityEngine;
