+++ Assets/MixedRealityToolkit/Providers/BaseDeviceManager.cs
@@ -9,7 +9,7 @@ using Microsoft.MixedReality.Toolkit.Core.Services;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.Core.Devices
+namespace Microsoft.MixedReality.Toolkit.Core.Providers 
 {
     /// <summary>
     /// Base Device manager to inherit from.
