+++ Assets/MixedRealityToolkit/Providers/BaseDeviceManager.cs.meta
