+++ Assets/MixedRealityToolkit/Providers/BaseSpatialObserver.cs.meta
