+++ Assets/MixedRealityToolkit/Providers/UnityInput/MouseDeviceManager.cs.meta
