+++ Assets/MixedRealityToolkit/Providers/UnityInput/UnityTouchDeviceManager.cs.meta
